# openapi.yaml
openapi: "3.1.0"
info:
  title: Lumora Privacy Coprocessor API
  version: "0.1.0"
  description: |
    HTTP JSON-RPC API for the Lumora privacy coprocessor.  
    All authenticated endpoints require the `X-API-Key` header when the
    `LUMORA_API_KEY` environment variable is set on the server.
  license:
    name: MIT OR Apache-2.0
servers:
  - url: http://127.0.0.1:3030
    description: Local development server

# ── Paths ─────────────────────────────────────────────────────────────
paths:
...
  /v1/deposit:
    post:
      summary: Deposit funds into the privacy pool
      operationId: deposit
      tags: [Write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositReq"
      responses:
        "200":
          description: Deposit accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositResp"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/transfer:
    post:
      summary: Submit a private transfer proof
      operationId: transfer
      tags: [Write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferReq"
      responses:
        "200":
          description: Transfer accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResp"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ContractError"
  /v1/withdraw:
    post:
      summary: Withdraw funds from the privacy pool
      operationId: withdraw
      tags: [Write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawReq"
      responses:
        "200":
          description: Withdrawal accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawResp"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ContractError"
...
  /v1/history:
    post:
      summary: Paginated event history
      operationId: history
      tags: [Read]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HistoryReq"
      responses:
        "200":
          description: Event history page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryResp"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResp"
    ContractError:
      description: Contract execution error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResp"
  schemas:
...
    DepositReq:
      type: object
      required: [commitment, amount]
      properties:
        commitment:
          type: string
          description: Hex-encoded note commitment (32-byte field element)
        amount:
          type: integer
          format: uint64
          description: Amount to deposit (in base units)
        asset:
          type: integer
          format: uint64
          default: 0
          description: "Asset identifier (0 = native token)"
    TransferReq:
      type: object
      required: [proof, merkle_root, nullifiers, output_commitments]
      properties:
        proof:
          type: string
          description: Hex-encoded proof bytes
        merkle_root:
          type: string
          description: Hex-encoded Merkle root
        nullifiers:
          type: array
          items:
            type: string
          minItems: 2
          maxItems: 2
          description: Hex-encoded nullifiers (exactly 2)
        output_commitments:
          type: array
          items:
            type: string
          minItems: 2
          maxItems: 2
          description: Hex-encoded output commitments (exactly 2)
        domain_chain_id:
          type: integer
          format: uint64
          description: Optional domain chain ID for V2 domain-separated nullifiers
        domain_app_id:
          type: integer
          format: uint64
          description: Optional domain application ID for V2 domain-separated nullifiers
    WithdrawReq:
      type: object
      required:
        [proof, merkle_root, nullifiers, output_commitments, amount, recipient]
      properties:
        proof:
          type: string
          description: Hex-encoded proof bytes
        merkle_root:
          type: string
          description: Hex-encoded Merkle root
        nullifiers:
          type: array
          items:
            type: string
          minItems: 2
          maxItems: 2
        output_commitments:
          type: array
          items:
            type: string
          minItems: 2
          maxItems: 2
        amount:
          type: integer
          format: uint64
        recipient:
          type: string
          description: Hex-encoded 32-byte recipient address
        asset:
          type: integer
          format: uint64
          default: 0
        domain_chain_id:
          type: integer
          format: uint64
        domain_app_id:
          type: integer
          format: uint64
...
    HistoryReq:
      type: object
      properties:
        offset:
          type: integer
          format: uint64
          default: 0
        limit:
          type: integer
          format: uint64
          default: 100
...
    DepositResp:
      type: object
      properties:
        leaf_index:
          type: integer
          format: uint64
        new_root:
          type: string
    TransferResp:
      type: object
      properties:
        leaf_indices:
          type: array
          items:
            type: integer
            format: uint64
          minItems: 2
          maxItems: 2
        new_root:
          type: string
    WithdrawResp:
      type: object
      properties:
        change_leaf_indices:
          type: array
          items:
            type: integer
            format: uint64
          minItems: 2
          maxItems: 2
        new_root:
          type: string
        amount:
          type: integer
          format: uint64
...
    HistoryResp:
      type: object
      properties:
        total:
          type: integer
          format: uint64
        events:
          type: array
          items:
            type: object
...
    ErrorResp:
      type: object
      properties:
        error:
          type: string
security:
  - ApiKeyAuth: []